refactor: update reducers types and redefine IRootState
