+++ src/store/reducers/header.ts
@@ -1,10 +1,13 @@
+import {Reducer} from 'redux';
+
 const SET_HEADER = 'SET_HEADER';
 
+type IHeaderAction = ReturnType<typeof setHeader>;
 export type HeaderItemType = {text: string; link?: string};
 
 const initialState: HeaderItemType[] = [];
 
-const header = function (state = initialState, action: ReturnType<typeof setHeader>) {
+const header: Reducer<HeaderItemType[], IHeaderAction> = function (state = initialState, action) {
     switch (action.type) {
         case SET_HEADER:
             return action.data;
