+++ src/store/reducers/singleClusterMode.ts
@@ -1,0 +1,5 @@
+function singleClusterMode(state = true) {
+    return state;
+}
+
+export default singleClusterMode;
