--- src/types/store/index.ts
@@ -1,4 +1,0 @@
-import {IDescribeRootStateSlice} from './describe';
-import {IHealthcheckInfoRootStateSlice} from './healthcheck';
-
-export interface IRootState extends IHealthcheckInfoRootStateSlice, IDescribeRootStateSlice {}
